# openapi/custom.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Maybe_Scalars-at-String_91_output_93__": {
				"type": "string",
				"nullable": true
			},
			"Maybe_Scalars-at-DateTime_91_output_93__": {
				"anyOf": [
					{
						"type": "string",
						"format": "date-time"
					},
					{
						"type": "string"
					}
				],
				"nullable": true
			},
...
			"Maybe_LinkBookModuleBookAuthorModuleAuthor_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/LinkBookModuleBookAuthorModuleAuthor"
					}
				],
				"nullable": true
			},
			"Maybe_Scalars-at-Int_91_output_93__": {
				"type": "number",
				"format": "double",
				"nullable": true
			},
			"Record_string.unknown_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"Maybe_Scalars-at-JSON_91_output_93__": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Record_string.unknown_"
					}
				],
				"nullable": true
			},
			"Maybe_Author_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Author"
					}
				],
				"nullable": true
			},
...
			"Maybe_LinkBookModuleBookLanguageModuleLanguage_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/LinkBookModuleBookLanguageModuleLanguage"
					}
				],
				"nullable": true
			},
			"Maybe_Book_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Book"
					}
				],
				"nullable": true
			},
...
			"Maybe_Language_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Language"
					}
				],
				"nullable": true
			},
			"ProductStatus": {
				"enum": [
					"draft",
					"proposed",
					"published",
					"rejected"
				],
				"type": "string"
			},
			"Maybe_Scalars-at-Float_91_output_93__": {
				"type": "number",
				"format": "double",
				"nullable": true
			},
			"Maybe_Product_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Product"
					}
				],
				"nullable": true
			},
			"Maybe_Array_Maybe_Product___": {
				"items": {
					"$ref": "#/components/schemas/Maybe_Product_"
				},
				"type": "array",
				"nullable": true
			},
			"ProductCollection": {
				"properties": {
					"products": {
						"$ref": "#/components/schemas/Maybe_Array_Maybe_Product___"
					},
					"deleted_at": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-DateTime_91_output_93__"
					},
					"updated_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"created_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"metadata": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-JSON_91_output_93__"
					},
					"handle": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"__typename": {
						"type": "string",
						"enum": [
							"ProductCollection"
						],
						"nullable": false
					}
				},
				"required": [
					"updated_at",
					"created_at",
					"handle",
					"title",
					"id"
				],
				"type": "object"
			},
			"Maybe_ProductCollection_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/ProductCollection"
					}
				],
				"nullable": true
			},
			"Maybe_ProductCategory_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/ProductCategory"
					}
				],
				"nullable": true
			},
			"ProductCategory": {
				"properties": {
					"deleted_at": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-DateTime_91_output_93__"
					},
					"updated_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"created_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"products": {
						"items": {
							"$ref": "#/components/schemas/Product"
						},
						"type": "array"
					},
					"category_children": {
						"items": {
							"$ref": "#/components/schemas/ProductCategory"
						},
						"type": "array"
					},
					"parent_category_id": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"parent_category": {
						"$ref": "#/components/schemas/Maybe_ProductCategory_"
					},
					"metadata": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-JSON_91_output_93__"
					},
					"rank": {
						"type": "number",
						"format": "double"
					},
					"is_internal": {
						"type": "boolean"
					},
					"is_active": {
						"type": "boolean"
					},
					"handle": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"__typename": {
						"type": "string",
						"enum": [
							"ProductCategory"
						],
						"nullable": false
					}
				},
				"required": [
					"updated_at",
					"created_at",
					"products",
					"category_children",
					"rank",
					"is_internal",
					"is_active",
					"handle",
					"description",
					"name",
					"id"
				],
				"type": "object"
			},
			"Product": {
				"properties": {
					"book": {
						"$ref": "#/components/schemas/Maybe_Book_"
					},
					"book_link": {
						"$ref": "#/components/schemas/Maybe_LinkProductProductBookModuleBook_"
					},
					"sales_channels": {
						"$ref": "#/components/schemas/Maybe_Array_Maybe_SalesChannel___"
					},
					"sales_channels_link": {
						"$ref": "#/components/schemas/Maybe_Array_Maybe_LinkProductSalesChannel___"
					},
					"metadata": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-JSON_91_output_93__"
					},
					"deleted_at": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-DateTime_91_output_93__"
					},
					"updated_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"created_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"external_id": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"discountable": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-Boolean_91_output_93__"
					},
					"images": {
						"items": {
							"$ref": "#/components/schemas/ProductImage"
						},
						"type": "array"
					},
					"options": {
						"items": {
							"$ref": "#/components/schemas/ProductOption"
						},
						"type": "array"
					},
					"variants": {
						"items": {
							"$ref": "#/components/schemas/ProductVariant"
						},
						"type": "array"
					},
					"tags": {
						"items": {
							"$ref": "#/components/schemas/ProductTag"
						},
						"type": "array"
					},
					"type_id": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"type": {
						"$ref": "#/components/schemas/Maybe_ProductType_"
					},
					"categories": {
						"$ref": "#/components/schemas/Maybe_Array_Maybe_ProductCategory___"
					},
					"collection_id": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"collection": {
						"$ref": "#/components/schemas/Maybe_ProductCollection_"
					},
					"material": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"mid_code": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"hs_code": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"origin_country": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"height": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-Float_91_output_93__"
					},
					"length": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-Float_91_output_93__"
					},
					"weight": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-Float_91_output_93__"
					},
					"width": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-Float_91_output_93__"
					},
					"thumbnail": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"status": {
						"$ref": "#/components/schemas/ProductStatus"
					},
					"is_giftcard": {
						"type": "boolean"
					},
					"description": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"subtitle": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"handle": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"__typename": {
						"type": "string",
						"enum": [
							"Product"
						],
						"nullable": false
					}
				},
				"required": [
					"updated_at",
					"created_at",
					"images",
					"options",
					"variants",
					"tags",
					"status",
					"is_giftcard",
					"handle",
					"title",
					"id"
				],
				"type": "object"
			},
			"Maybe_Array_Maybe_ProductCategory___": {
				"items": {
					"$ref": "#/components/schemas/Maybe_ProductCategory_"
				},
				"type": "array",
				"nullable": true
			},
			"ProductType": {
				"properties": {
					"deleted_at": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-DateTime_91_output_93__"
					},
					"updated_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"created_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"metadata": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-JSON_91_output_93__"
					},
					"value": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"__typename": {
						"type": "string",
						"enum": [
							"ProductType"
						],
						"nullable": false
					}
				},
				"required": [
					"updated_at",
					"created_at",
					"value",
					"id"
				],
				"type": "object"
			},
			"Maybe_ProductType_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/ProductType"
					}
				],
				"nullable": true
			},
			"ProductTag": {
				"properties": {
					"products": {
						"$ref": "#/components/schemas/Maybe_Array_Maybe_Product___"
					},
					"metadata": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-JSON_91_output_93__"
					},
					"value": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"__typename": {
						"type": "string",
						"enum": [
							"ProductTag"
						],
						"nullable": false
					}
				},
				"required": [
					"value",
					"id"
				],
				"type": "object"
			},
			"ProductOptionValue": {
				"properties": {
					"deleted_at": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-DateTime_91_output_93__"
					},
					"updated_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"created_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"metadata": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-JSON_91_output_93__"
					},
					"option_id": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"option": {
						"$ref": "#/components/schemas/Maybe_ProductOption_"
					},
					"value": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"__typename": {
						"type": "string",
						"enum": [
							"ProductOptionValue"
						],
						"nullable": false
					}
				},
				"required": [
					"updated_at",
					"created_at",
					"value",
					"id"
				],
				"type": "object"
			},
			"ProductOption": {
				"properties": {
					"deleted_at": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-DateTime_91_output_93__"
					},
					"updated_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"created_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"metadata": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-JSON_91_output_93__"
					},
					"values": {
						"items": {
							"$ref": "#/components/schemas/ProductOptionValue"
						},
						"type": "array"
					},
					"product_id": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"product": {
						"$ref": "#/components/schemas/Maybe_Product_"
					},
					"title": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"__typename": {
						"type": "string",
						"enum": [
							"ProductOption"
						],
						"nullable": false
					}
				},
				"required": [
					"updated_at",
					"created_at",
					"values",
					"title",
					"id"
				],
				"type": "object"
			},
			"Maybe_ProductOption_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/ProductOption"
					}
				],
				"nullable": true
			},
...
			"Maybe_LinkProductVariantInventoryItem_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/LinkProductVariantInventoryItem"
					}
				],
				"nullable": true
			},
			"Maybe_Array_Maybe_LinkProductVariantInventoryItem___": {
				"items": {
					"$ref": "#/components/schemas/Maybe_LinkProductVariantInventoryItem_"
				},
				"type": "array",
				"nullable": true
			},
...
			"Maybe_Array_Maybe_StockLocation___": {
				"items": {
					"$ref": "#/components/schemas/Maybe_StockLocation_"
				},
				"type": "array",
				"nullable": true
			},
...
			"Maybe_Scalars-at-Boolean_91_output_93__": {
				"type": "boolean",
				"nullable": true
			},
...
			"Maybe_OrderLineItem_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/OrderLineItem"
					}
				],
				"nullable": true
			},
...
			"ProductVariant": {
				"properties": {
					"order_items": {
						"$ref": "#/components/schemas/Maybe_Array_Maybe_OrderLineItem___"
					},
					"price_set": {
						"$ref": "#/components/schemas/Maybe_PriceSet_"
					},
					"price_set_link": {
						"$ref": "#/components/schemas/Maybe_LinkProductVariantPriceSet_"
					},
					"inventory": {
						"$ref": "#/components/schemas/Maybe_Array_Maybe_InventoryItem___"
					},
					"inventory_items": {
						"$ref": "#/components/schemas/Maybe_Array_Maybe_LinkProductVariantInventoryItem___"
					},
					"deleted_at": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-DateTime_91_output_93__"
					},
					"updated_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"created_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"variant_rank": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-Int_91_output_93__"
					},
					"product_id": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"product": {
						"$ref": "#/components/schemas/Maybe_Product_"
					},
					"metadata": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-JSON_91_output_93__"
					},
					"options": {
						"items": {
							"$ref": "#/components/schemas/ProductOptionValue"
						},
						"type": "array"
					},
					"width": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-Float_91_output_93__"
					},
					"height": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-Float_91_output_93__"
					},
					"length": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-Float_91_output_93__"
					},
					"weight": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-Float_91_output_93__"
					},
					"material": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"mid_code": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"origin_country": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"hs_code": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"requires_shipping": {
						"type": "boolean"
					},
					"manage_inventory": {
						"type": "boolean"
					},
					"allow_backorder": {
						"type": "boolean"
					},
					"upc": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"ean": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"barcode": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"sku": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"title": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"__typename": {
						"type": "string",
						"enum": [
							"ProductVariant"
						],
						"nullable": false
					}
				},
				"required": [
					"updated_at",
					"created_at",
					"options",
					"requires_shipping",
					"manage_inventory",
					"allow_backorder",
					"title",
					"id"
				],
				"type": "object"
			},
...
			"Maybe_Array_Maybe_OrderLineItem___": {
				"items": {
					"$ref": "#/components/schemas/Maybe_OrderLineItem_"
				},
				"type": "array",
				"nullable": true
			},
...
			"Maybe_SalesChannel_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/SalesChannel"
					}
				],
				"nullable": true
			},
			"LinkProductSalesChannel": {
				"properties": {
					"deletedAt": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"updatedAt": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"sales_channel": {
						"$ref": "#/components/schemas/Maybe_SalesChannel_"
					},
					"product": {
						"$ref": "#/components/schemas/Maybe_Product_"
					},
					"sales_channel_id": {
						"type": "string"
					},
					"product_id": {
						"type": "string"
					},
					"__typename": {
						"type": "string",
						"enum": [
							"LinkProductSalesChannel"
						],
						"nullable": false
					}
				},
				"required": [
					"updatedAt",
					"createdAt",
					"sales_channel_id",
					"product_id"
				],
				"type": "object"
			},
			"Maybe_LinkProductSalesChannel_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/LinkProductSalesChannel"
					}
				],
				"nullable": true
			},
			"Maybe_Array_Maybe_LinkProductSalesChannel___": {
				"items": {
					"$ref": "#/components/schemas/Maybe_LinkProductSalesChannel_"
				},
				"type": "array",
				"nullable": true
			},
...
			"Maybe_Array_Maybe_LinkSalesChannelStockLocation___": {
				"items": {
					"$ref": "#/components/schemas/Maybe_LinkSalesChannelStockLocation_"
				},
				"type": "array",
				"nullable": true
			},
			"SalesChannel": {
				"properties": {
					"stock_locations": {
						"$ref": "#/components/schemas/Maybe_Array_Maybe_StockLocation___"
					},
					"locations_link": {
						"$ref": "#/components/schemas/Maybe_Array_Maybe_LinkSalesChannelStockLocation___"
					},
					"orders": {
						"$ref": "#/components/schemas/Maybe_Array_Maybe_Order___"
					},
					"api_keys_link": {
						"$ref": "#/components/schemas/Maybe_Array_Maybe_LinkPublishableApiKeySalesChannel___"
					},
					"products_link": {
						"$ref": "#/components/schemas/Maybe_Array_Maybe_LinkProductSalesChannel___"
					},
					"deleted_at": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-DateTime_91_output_93__"
					},
					"updated_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"metadata": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-JSON_91_output_93__"
					},
					"created_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"is_disabled": {
						"type": "boolean"
					},
					"description": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"__typename": {
						"type": "string",
						"enum": [
							"SalesChannel"
						],
						"nullable": false
					}
				},
				"required": [
					"updated_at",
					"created_at",
					"is_disabled",
					"name",
					"id"
				],
				"type": "object"
			},
...
			"Maybe_Array_Maybe_SalesChannel___": {
				"items": {
					"$ref": "#/components/schemas/Maybe_SalesChannel_"
				},
				"type": "array",
				"nullable": true
			},
...
			"Maybe_InventoryItem_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/InventoryItem"
					}
				],
				"nullable": true
			},
			"Maybe_Array_Maybe_InventoryItem___": {
				"items": {
					"$ref": "#/components/schemas/Maybe_InventoryItem_"
				},
				"type": "array",
				"nullable": true
			},
			"LinkProductVariantPriceSet": {
				"properties": {
					"deletedAt": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"updatedAt": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"variant": {
						"$ref": "#/components/schemas/Maybe_Product_"
					},
					"price_set_id": {
						"type": "string"
					},
					"variant_id": {
						"type": "string"
					},
					"__typename": {
						"type": "string",
						"enum": [
							"LinkProductVariantPriceSet"
						],
						"nullable": false
					}
				},
				"required": [
					"updatedAt",
					"createdAt",
					"price_set_id",
					"variant_id"
				],
				"type": "object"
			},
			"Maybe_LinkProductVariantPriceSet_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/LinkProductVariantPriceSet"
					}
				],
				"nullable": true
			},
			"Maybe_PriceSet_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/PriceSet"
					}
				],
				"nullable": true
			},
...
			"PriceSet": {
				"properties": {
					"shipping_option": {
						"$ref": "#/components/schemas/Maybe_ShippingOption_"
					},
					"shipping_option_link": {
						"$ref": "#/components/schemas/Maybe_LinkShippingOptionPriceSet_"
					},
					"variant": {
						"$ref": "#/components/schemas/Maybe_ProductVariant_"
					},
					"variant_link": {
						"$ref": "#/components/schemas/Maybe_LinkProductVariantPriceSet_"
					},
					"calculated_price": {
						"$ref": "#/components/schemas/Maybe_CalculatedPriceSet_"
					},
					"prices": {
						"$ref": "#/components/schemas/Maybe_Array_Maybe_Price___"
					},
					"id": {
						"type": "string"
					},
					"__typename": {
						"type": "string",
						"enum": [
							"PriceSet"
						],
						"nullable": false
					}
				},
				"required": [
					"id"
				],
				"type": "object"
			},
			"ProductImage": {
				"properties": {
					"deleted_at": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-DateTime_91_output_93__"
					},
					"updated_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"created_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"metadata": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-JSON_91_output_93__"
					},
					"url": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"__typename": {
						"type": "string",
						"enum": [
							"ProductImage"
						],
						"nullable": false
					}
				},
				"required": [
					"updated_at",
					"created_at",
					"url",
					"id"
				],
				"type": "object"
			},
			"Maybe_LinkProductProductBookModuleBook_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/LinkProductProductBookModuleBook"
					}
				],
				"nullable": true
			},
			"LinkProductProductBookModuleBook": {
				"properties": {
					"deletedAt": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"updatedAt": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"product": {
						"$ref": "#/components/schemas/Maybe_Product_"
					},
					"book_id": {
						"type": "string"
					},
					"product_id": {
						"type": "string"
					},
					"__typename": {
						"type": "string",
						"enum": [
							"LinkProductProductBookModuleBook"
						],
						"nullable": false
					}
				},
				"required": [
					"updatedAt",
					"createdAt",
					"book_id",
					"product_id"
				],
				"type": "object"
			},
...
			"Maybe_LinkQuoteModuleQuoteBookModuleBook_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/LinkQuoteModuleQuoteBookModuleBook"
					}
				],
				"nullable": true
			},
			"Quote": {
				"properties": {
					"book": {
						"$ref": "#/components/schemas/Maybe_Book_"
					},
					"book_link": {
						"$ref": "#/components/schemas/Maybe_LinkQuoteModuleQuoteBookModuleBook_"
					},
					"author": {
						"$ref": "#/components/schemas/Maybe_Author_"
					},
					"author_link": {
						"$ref": "#/components/schemas/Maybe_LinkQuoteModuleQuoteAuthorModuleAuthor_"
					},
					"deleted_at": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-DateTime_91_output_93__"
					},
					"updated_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"created_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"details": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"content": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"__typename": {
						"type": "string",
						"enum": [
							"Quote"
						],
						"nullable": false
					}
				},
				"required": [
					"updated_at",
					"created_at",
					"content",
					"id"
				],
				"type": "object"
			},
			"Maybe_Quote_": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Quote"
					}
				],
				"nullable": true
			},
			"Book": {
				"properties": {
					"quote": {
						"$ref": "#/components/schemas/Maybe_Quote_"
					},
					"quote_link": {
						"$ref": "#/components/schemas/Maybe_LinkQuoteModuleQuoteBookModuleBook_"
					},
					"product": {
						"$ref": "#/components/schemas/Maybe_Product_"
					},
					"product_link": {
						"$ref": "#/components/schemas/Maybe_LinkProductProductBookModuleBook_"
					},
					"language": {
						"$ref": "#/components/schemas/Maybe_Language_"
					},
					"language_link": {
						"$ref": "#/components/schemas/Maybe_LinkBookModuleBookLanguageModuleLanguage_"
					},
					"author": {
						"$ref": "#/components/schemas/Maybe_Author_"
					},
					"author_link": {
						"$ref": "#/components/schemas/Maybe_LinkBookModuleBookAuthorModuleAuthor_"
					},
					"deleted_at": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-DateTime_91_output_93__"
					},
					"updated_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"created_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"raw_publishDate": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-JSON_91_output_93__"
					},
					"publisher": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"publishDate": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"pagesCount": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-Int_91_output_93__"
					},
					"isbn": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"ebookVersionEbook": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"dimensions": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"audioVersionUrl": {
						"$ref": "#/components/schemas/Maybe_Scalars-at-String_91_output_93__"
					},
					"id": {
						"type": "string"
					},
					"__typename": {
						"type": "string",
						"enum": [
							"Book"
						],
						"nullable": false
					}
				},
				"required": [
					"updated_at",
					"created_at",
					"id"
				],
				"type": "object"
			},
...
			"TApiPaginatedReponse_Product-Array_": {
				"properties": {
					"offset": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"count": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Product"
						},
						"type": "array"
					}
				},
				"required": [
					"offset",
					"limit",
					"count",
					"data"
				],
				"type": "object"
			},
			"TApiDataReponse_Product_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/Product"
					}
				},
				"required": [
					"data"
				],
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@ztf-library/code",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/store/books": {
			"get": {
				"operationId": "GetBooks",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TApiPaginatedReponse_Product-Array_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					}
				},
				"description": "Get books, paginated using optional limit and offset",
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/store/books/category/{category_id}": {
			"get": {
				"operationId": "GetBooksByCategory",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TApiPaginatedReponse_Product-Array_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					}
				},
				"description": "Get books from a category, paginated using optional limit and offset",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "category_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/store/books/{slug}": {
			"get": {
				"operationId": "GetBookBySlug",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TApiDataReponse_Product_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					}
				},
				"description": "Get a book by its slug",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "slug",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}